Rewrite language get_terms()
order of the languages terms, using the classic usort() function resulting in a simpler code and hopefully a stable sort.

New Code (with load-nav-menus.php)
Original: "Hey WordPress, add this box to the Menus screen!" (Shouted on every page load).
New Code: "Wait... are we on the Menus screen? Yes? Okay, now add this box." (Polite and efficient).